+++ packages/rstream-csp/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.116](https://github.com/thi-ng/umbrella/compare/@thi.ng/rstream-csp@0.1.115...@thi.ng/rstream-csp@0.1.116) (2018-11-24)
+
+**Note:** Version bump only for package @thi.ng/rstream-csp
+
+
+
+
+
 ## [0.1.115](https://github.com/thi-ng/umbrella/compare/@thi.ng/rstream-csp@0.1.114...@thi.ng/rstream-csp@0.1.115) (2018-11-08)
 
 **Note:** Version bump only for package @thi.ng/rstream-csp