+++ packages/rstream-graph/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.41](https://github.com/thi-ng/umbrella/compare/@thi.ng/rstream-graph@2.1.40...@thi.ng/rstream-graph@2.1.41) (2018-11-24)
+
+**Note:** Version bump only for package @thi.ng/rstream-graph
+
+
+
+
+
 ## [2.1.40](https://github.com/thi-ng/umbrella/compare/@thi.ng/rstream-graph@2.1.39...@thi.ng/rstream-graph@2.1.40) (2018-11-08)
 
 **Note:** Version bump only for package @thi.ng/rstream-graph