+++ packages/rstream-query/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.54](https://github.com/thi-ng/umbrella/compare/@thi.ng/rstream-query@0.3.53...@thi.ng/rstream-query@0.3.54) (2018-11-24)
+
+**Note:** Version bump only for package @thi.ng/rstream-query
+
+
+
+
+
 ## [0.3.53](https://github.com/thi-ng/umbrella/compare/@thi.ng/rstream-query@0.3.52...@thi.ng/rstream-query@0.3.53) (2018-11-08)
 
 **Note:** Version bump only for package @thi.ng/rstream-query